backend/backend/requirements.txt
openai-whisper
ffmpeg
setuptools-rust